renamed:    database/migrations/2025_02_13_023310_create_violations_table.php -> database/migrations/2025_03_11_055624_create_violations_table.php 	renamed:    database/migrations/2025_02_13_023310_create_video_violations_table.php -> database/migrations/2025_03_11_060330_create_video_violations_table.php

+++ database/migrations/2025_03_11_060330_create_video_violations_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::create('video_violations', function (Blueprint $table) {
             $table->id(); // Primary Key
             $table->uuid('video_id');
-            $table->unsignedTinyInteger('violation_id'); // Match violations.id
+            $table->unsignedBigInteger('violation_id'); // Updated to match violations.id
             $table->timestamps();
         
             // Foreign Key Constraints
@@ -21,7 +21,6 @@ return new class extends Migration {
             $table->foreign('violation_id')->references('id')->on('violations')->onDelete('cascade');
         });        
     }
-
 
     /**
      * Reverse the migrations.
